# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scheduling Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scheduling
    description: Scheduling operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /state:
    get:
      operationId: getState
      tags: [scheduling]
      summary: Get full scheduling state
      responses:
        "200":
          description: Current scheduling state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulingState"
    put:
      operationId: setState
      tags: [scheduling]
      summary: Save full scheduling state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulingState"
      responses:
        "200":
          description: Saved state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulingState"

  /schedule/auto-assign:
    post:
      operationId: autoAssign
      tags: [scheduling]
      summary: Auto-assign schedule for a month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoAssignRequest"
      responses:
        "200":
          description: Updated state after auto-assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulingState"

  /summary:
    get:
      operationId: getSummary
      tags: [scheduling]
      summary: Get summary stats for a month
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ShiftType:
      type: string
      enum: [FHD, BHD, "Part Time", Vacation]

    Associate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        shiftType:
          $ref: "#/components/schemas/ShiftType"
        active:
          type: boolean
      required: [id, name, shiftType, active]

    PoolRule:
      type: object
      properties:
        sunWed:
          type: boolean
        wedSat:
          type: boolean
        partTime:
          type: boolean
        skip:
          type: boolean
      required: [sunWed, wedSat, partTime, skip]

    Assignment:
      type: object
      properties:
        mainId:
          type: string
        supportId:
          type: string
        categoryMain:
          $ref: "#/components/schemas/ShiftType"
        categorySupport:
          $ref: "#/components/schemas/ShiftType"
      required: [mainId, supportId, categoryMain, categorySupport]

    Backup:
      type: object
      properties:
        mainId:
          type: string
        backupId:
          type: string
      required: [mainId, backupId]

    SchedulingState:
      type: object
      properties:
        associates:
          type: array
          items:
            $ref: "#/components/schemas/Associate"
        pooling:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PoolRule"
        assignments:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Assignment"
        backups:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Backup"
      required: [associates, pooling, assignments, backups]

    AutoAssignRequest:
      type: object
      properties:
        month:
          type: string
          pattern: '^\d{4}-\d{2}$'
        overwrite:
          type: boolean
          default: false
      required: [month]

    AssociateSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        shiftType:
          $ref: "#/components/schemas/ShiftType"
        mainCount:
          type: integer
        supportCount:
          type: integer
        backupCount:
          type: integer
      required: [id, name, shiftType, mainCount, supportCount, backupCount]

    ScheduleSummary:
      type: object
      properties:
        month:
          type: string
        totalDays:
          type: integer
        assignedDays:
          type: integer
        unassignedDays:
          type: integer
        associates:
          type: array
          items:
            $ref: "#/components/schemas/AssociateSummary"
      required: [month, totalDays, assignedDays, unassignedDays, associates]